scripts/requirements.txt
opencv-python
transformers
torch
pillow
python-osc
